+++ RELEASES.md
@@ -3,6 +3,7 @@
 Released on TBD (UTC).
 
 ### Enhancements
+- Added optional `depth` param for some `OrderBook` methods
 - Refactored to use `exchange` MIC code as `venue` for instrument IDs with Databento GLBX dataset (#2108), thanks @faysou
 - Refactored to use `self.config` attributes consistently (#2120), thanks @stefansimik